Not inspired
I forgot the fact that I still have no clear way
to detect if the UI can be run or not.

But the biggest issue is that I'm clearly not
inspired on this project.
I still have no clear idea where this is going,
even though there's a roadmap.

QT is sucking out the will I'd like to put on this
project, due to a FUCKING HORRENDOUS toolset.
The more I'm using it, the more I feel that QML is
just "that garbage that QT doesn't know what to do
with". This reminds that I'll have to make a converter
from Godot Engine, or whatever over UI maker that
makes a bit of sense in this era.

So yeah...

The best way to tackle this would be to have a UI
focused on presenting all the available docker images,
while also providing extended view and toolset for
installed images.
This can be done. However, then I'm still wondering
how to generalize this afterwards.
Hence the 'I feel I'm alone on this project that is
not mine'.

So, unless there's some real movement on the project,
this might be one of the last commit.

+++ testQT.py
@@ -2,6 +2,7 @@ import logging
 import os
 from pathlib import Path
 import shutil
+import subprocess
 import sys
 
 from argparse import ArgumentParser, RawDescriptionHelpFormatter
@@ -10,7 +11,7 @@ from renderers.qt.widgets import SelectionMenu, provide_user_choice
 logging.basicConfig(level=logging.DEBUG)
 
 config = {
-    'templates_dir': './softwares',
+    'templates_dir': '/usr/share/armbian/configurator/modules/softwares',
     'install_dir': '/opt/armbian/docker',
     'install': [],
     'remove': []
@@ -40,11 +41,20 @@ def list_dirs(from_dir=".", containing_files=[]) -> list:
 class Module:
     def __init__(self, configuration):
         self.configuration = configuration
+        configuration["install_dir"] = Path(configuration("install_dir"))
+        configuration["templates_dir"] = Path(configuration["templates_dir"))
+
+    def installed_path(self, service_name:str):
+        service_path = (configuration["install_dir"] / service_name)
+
 
     def docker_install(self, configuration=None):
         configuration = self.configuration if configuration == None else configuration
-        install_path = Path(configuration["install_dir"])
-        templates_path = Path(configuration["templates_dir"])
+        if len(configuration["install"]) == 0:
+            print('No docker to install')
+            return
+        install_path = configuration["install_dir"]
+        templates_path = configuration["templates_dir"]
         try:
             install_path.mkdir( mode=0o755, parents=True, exist_ok=True )
         except Exception as e:
@@ -58,22 +68,43 @@ class Module:
 
     def docker_remove(self, configuration=None):
         configuration = self.configuration if configuration == None else configuration
-        install_path = Path(configuration["install_dir"])
+        if len(configuration["remove"]) is 0:
+            print('No docker to remove')
+            return
 
-        for dir_to_remove in configuration["remove"]:
-            remove_path = install_path / dir_to_remove
-            try:
-                shutil.rmtree(remove_path)
-            except Exception as e:
-                print(f'Could not delete {remove_path}')
+        for docker_compose_service in configuration["remove"]:
+            remove_path = self.installed_path(docker_compose_service)
+            if remove_path.exists():
+                try:
+                    shutil.rmtree(remove_path)
+                except Exception as e:
+                    print(f'Could not delete {remove_path}')
+                    continue
+            else:
+                print(f'{remove_path} does not exist. Skipping {docker_compose_servivce}')
                 continue
 
+    def start_script(self, script="", args=[]):
+        return subprocess.Popen([script] + args)
+
+    def start_service(self, docker_service_name):
+        service_path = self.installed_path(docker_service_name)
+        if service_path.exists():
+            self.start_script('./scripts/start_service.sh', [service_path])
+
+    def stop_service(self, docker_service_name):
+        service_path = self.installed_path(docker_service_name)
+        if service_path.exists():
+            self.start_script('./scripts/stop_service.sh', [service_path])
+
     def run(self):
+
         self.docker_remove()
         self.docker_install()
-        
+
 def module_payload(configuration):
     docker_install(configuration["install"])
+
 
 if __name__ == '__main__':
     # TODO : Parse from configuration file
@@ -82,6 +113,9 @@ if __name__ == '__main__':
     images_list = '\n\t' + '\n\t'.join(available_dockers)
     parser = ArgumentParser(description='Armbian docker installation module', formatter_class=RawDescriptionHelpFormatter, epilog='IMAGES :' + images_list)
     parser.add_argument('--install', nargs='+', metavar='IMAGE', choices=available_dockers, help='Add docker installations')
+    parser.add_argument('--start', nargs='+', metavar='IMAGE', choices=installed_dockers, help='Start selected installed docker-compose images')
+    parser.add_argument('--stop', nargs='+', metavar='IMAGE', choices=installed_dockers, help='Stop selected installed docker-compose images')
+    parser;add_argument('--edit', nargs=1, metavar='IMAGE', choices=installed_dockers, help='Edit an installed docker-compose image YAML file')
     parser.add_argument('--templates-dir', nargs='?', default=config['templates_dir'], help='Directory path to list templates from')
     parser.add_argument('--install-dir', nargs='?', default=config['install_dir'], help='Directory path where docker images are installed')
     args = parser.parse_args()
@@ -90,7 +124,7 @@ if __name__ == '__main__':
     templates_dir = args.templates_dir
     
     if args.install == None or len(args.install) == 0:
-        #try:
+        try:
             installed_set = set(installed_dockers)
             selected, unselected = provide_user_choice(available_dockers, installed_set)
             available_set = set(available_dockers)
@@ -103,10 +137,7 @@ if __name__ == '__main__':
             config["remove"] = to_remove
             Module(config).run()
 
-        #except Exception as e:
-        #    print("An exception occured")
-        #    print(e)
-        #    print(args.install)
-        #    parser.print_help()
-
+        except Exception as e:
+            parser.print_help()
+            print(f'A problem araised {e}')
